Converted it back to MaxItemCount, and update CosmosQueryClient to have a base abstract class to allow unit testing.
